decision-engine/requirements.txt
confluent-kafka==2.3.0
httpx==0.27.0
psycopg2-binary==2.9.9
pyyaml==6.0.1
watchdog==4.0.0
websockets==13.1
structlog==25.5.0
